Fix: 소셜 로그인과 일반 로그인 상태 관리 분리
